favorites added, my recipes added, comments started

+++ src/main/java/ru/itis/foodbook_app/models/Favorite.java
@@ -19,7 +19,7 @@ public class Favorite {
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     protected Long id;
 
-    protected int userId;
-    protected int recipeId;
+    protected Long userId;
+    protected Long recipeId;
 
 }
